Seeding the SF.net repos with revision 742 of file:///Users/rolo/repos/zaphod/trunk/estar/nextgen
Revisions prior to this move are not expected to be relevant for a
wider community. They contain changes from the PhD version (simply
called "wfront") to the redesigns that led to the adoption of BGL for
C-space representation, introduction of "rhs" keys to replace
polymorphic events, migration to GNU autotools, and a plethora of
lesser changes.


git-svn-id: https://estar.svn.sourceforge.net/svnroot/estar/trunk/estar@2 255146f7-950e-0410-bd4e-f0acf59374a5