Added a foundation for developing growable grids. The idea is that the robot can start exploring with just a relatively small local grid, and expand it as more of the environment comes into view.
The idea is the following: Instead of opting for a general
multiresolution grid approach, I've decided to keep it relatively
simple and aim for a "grid of grids" where each subgrid (a cell of the
supergrid) contains a estar::Grid instance. Whenever we add a new
super-cell, we instantiate a new estar::Grid and "glue" its edges to
its surrounding neighbors. The supergrid will grow in a rectangular
fashion to keep it easy to manage, but some (or even most) of the
super-cells will just be null-pointers. The structure of the
sub-cells, which contain actual estar::Grid instances, will reflect
which regions of the supergrid have *actual* data in them.

The next step is to implement a Supergrid class and either:
1) make its API compatible with Grid to use it interchangeably
2) or change the current code such that it directly accesses Supergrid
3) or change the code such that it only needs to see a given sub-cell
   at a time

Which one of these is most appropriate will emerge during work on
Supergrid. The foundation for implementing it are coded as templates
that provide a type-neutral "automatically" growing 2D array:

estar/sdeque.hpp:
 Implementation of "deque" with signed index, such that (conceptually)
 the data stays in place when resizing the deque.

estar/flexgrid.hpp:
 Implementation of a "flexible" grid, which is like a sdeque in two
 dimensions.

bin/test_flexgrid.cpp:
  Test program for sdeque and flexgrid.


git-svn-id: https://estar.svn.sourceforge.net/svnroot/estar/trunk/estar@204 255146f7-950e-0410-bd4e-f0acf59374a5